refactor: move parseServerlessConfig to integrating buildIR
